Split up ReportTabs into `ReportTablist` and `ReportTabpanel`
