+++ src/components/ReportTabpanel.scss
@@ -1,0 +1,4 @@
+.isr-report-tabpanel {
+  min-height: 0;
+  flex: 1;
+}
