--- src/components/ReportTabs.scss
@@ -1,5 +1,0 @@
-.isr-report-tabs-wrapper,
-.isr-report-tabs-content {
-  min-height: 0;
-  flex: 1;
-}
