+++ src/components/index.tsx
@@ -2,6 +2,7 @@ export * from './Report';
 export * from './ReportTitle';
 export * from './ReportTimestamp';
 export * from './ReportBanner';
-export * from './ReportTabs';
+export * from './ReportTablist';
+export * from './ReportTabpanel';
 export * from './FilesTable';
 export * from './DetailsTable';
